current overlay canvas + inspector of rect tform

+++ Slice/SliceEditor/src/Inspector/ComponentPropertiesGUI.cpp
@@ -73,7 +73,7 @@ namespace SliceEditor
 		return changed;
 	}
 
-	bool DragIntInput(Registry& reg, const char* id, int& val, const char* format, float min, float max)
+	bool DragIntInput(Registry& reg, const char* id, int& val, const char* format, int min, int max)
 	{
 		static int oldVal{};
 
@@ -153,7 +153,7 @@ namespace SliceEditor
 		return changed;
 	}
 
-	bool DragIntInputHeader(Registry& reg, const char* property_label, const char* id, int& val, const char* format, float min, float max)
+	bool DragIntInputHeader(Registry& reg, const char* property_label, const char* id, int& val, const char* format, int min, int max)
 	{
 		bool changed = false;
 		ImGui::Text(property_label);
@@ -197,7 +197,7 @@ namespace SliceEditor
 		return changed;
 	}
 
-	bool DragIntInputScriptHeader(Registry& reg, std::function<void(std::string, int)> func, const char* property_label, const char* id, int& val, const char* format, float min, float max)
+	bool DragIntInputScriptHeader(Registry& reg, std::function<void(std::string, int)> func, const char* property_label, const char* id, int& val, const char* format, int min, int max)
 	{
 		ImGui::Text(property_label);
 		ImGui::SameLine(150.f);
